fix(go): indexed property access incorrectly renamed
Go caches translated names for identifiers on the value symbol, however
TypeScript since 3.9 changed how indexed property accesses are
represented and uses a single symbol for all accesses through an index
signature, causing the cache to incorrectly map all indexed accesses to
the name of the very first one.

Fixes this issue by not using the symbol as a cache key when it
corresponds to a transient signature.