fix(go): indexed property access incorrectly renamed (#7)
Go caches translated names for identifiers on the value symbol, however
TypeScript since 3.9 changed how indexed property accesses are
represented and uses a single symbol for all accesses through an index
signature, causing the cache to incorrectly map all indexed accesses to
the name of the very first one.

Fixes this issue by not using the symbol as a cache key when it
corresponds to a transient signature.



---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0