chore: replace mock-fs with memfs that works in node20 (#355)
We are affected by this `mock-fs` bug:
https://github.com/tschaub/mock-fs/issues/380
Seems like there is no activity on the repo recently, so we switch to a
new tool.

However `memfs` only works for the test case in `assemblies.test.ts`.
So we keep `mock-fs` for the `rosetta.test.ts` test cases.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0