chore: support backporting of PRs to other versions (#639)
With Mergify we got backporting of PRs to other branches for free. Since
this repo doesn't use Mergify, we have lost access to the feature.
However we still need to maintain a number of branches.

This change introduce a new PR backporting system based on labels.

Once this is working smoothly, we can backport (hehe) this to projen
upstream.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0