chore: untangle runtime from config code (#669)
In #530 we made a change to ReleasesDocument.
This was intended as a configuration change, but the interface resides
in runtime code. The change broke the automatic backporting of
configuration updates to maintenance branches, because we don't
auto-backport runtime code.

The solution is that we do not mix config and runtime code, so
backporting config will always be complete.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0