chore: make backport component backwards compatible
The PR backport component was implemented using a recent version of projen.
The automatic config backport runs BEFORE dependencies are upgrades.
That means we use the backported config with a not-updated version of projen.

In most cases this is fine, because the versions won't be very different.
However currently we are running behind projen upgrades, so the versions have diverged a lot, which in turn causes a type error.

This change avoids the type error to unblock the automatic config backport and dependency upgrades.
I have verified that this is really just type error and the code still works fine.