fix: 'TargetLanguage' has no exported member 'VISUALIZE' (#1238)
The signature overwrite for targetName using TargetLanguage.VISUALIZE
causes and error, because the enum value is stripped from the generate
type declarations. We need to also strip the signature overwrite.

I manually checked the output to confirm. Not sure how else to test
this.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0