fix: error TS2339: Property 'VISUALIZE' does not exist on type 'typeof TargetLanguage' (#1293)
We are using the `@internal` directive to instruct typescript compiler
to drop some statements during compilation.
This may cause the compiler to emit invalid declarations if we are not
careful.
To guarantee that our declarations are valid, we do an additional run of
tsc on the output.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0