chore: restore ability to release maintenance version lines (#1386)
Same as https://github.com/aws/jsii-compiler/pull/998, but for
jsii-rosetta

This repo was using an automatic backport feature for repository config.
As part of daily upgrades, the projen config would be checked out from
`main` and applied to maintenance branches. However because we also
defined the current version in this config, this inadvertently changed
the current version for maintenance branches as well! In practice that
meant that the release tagging workflow tagged the wrong release lines.

This fix does two things: 
- Disable the automatic backport of configuration in favor of the new,
explicit label based backport feature (see it in action on this PR!)
- Add an explicit release line option to the tagging workflow, that will
be compared against the detected TypeScript version and fail if they
don't match.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0