feat: TypeScript 5.5  (#1778)
Upgrade to TypeScript 5.5.

Closes #1769 

---

Due to the newly introduced [JSDoc @import
Tag](https://devblogs.microsoft.com/typescript/announcing-typescript-5-5/#the-jsdoc-import-tag)
feature, the signature for `analyzeImportDeclaration` has changed to
accept both an `ImportDeclaration` and a `JSDocImportTag` which are
compatible enough for the purpose of this function.

Existing tests are covering the code paths that call
`analyzeImportDeclaration`.
None of the code paths currently call `analyzeImportDeclaration` with a
`JSDocImportTag`, but I opted for this change instead of a guard to
future proof any calls. The other option would have been to guard each
of the calls, however that would impose additional knowledge on the
calling site that seemed unnecessary to dangerous to me.

Adds tests for the new jsdoc-import-tag feature, though they are
currently not treated any differently than other comments. In rosetta,
comments are mostly presumed to be pre-processed by jsii and in the
structured assembly.

Also added 3 translation test cases that were previously just missing.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0

---------

Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>