chore: release workflow won't release dev releases
Tagging can be done for pre-releases, as well as for official releases.
For the purposes of this PR body, we will call those different
styles of tag a "tag class".

The tagging workflow works this way:

1. If the current commit is already tagged, stop.
2. Fetch all tags.
3. If the current commit is already tagged with an equivalent or higher
  tag class, stop.

The upshot of this is that if the current commit is already tagged
as a `dev` tag, through check (1) it will not be tagged again
as an official release; we never reach check (3) which would have
correctly ignored pre-release tags.

It seems like check (3) fully subsumes check (1), so I'm removing
check (1).