feat: add support for class-covariant-overrides feature (#3350)
Add 'class-covariant-overrides' to the list of supported assembly
features to enable processing of assemblies that use this jsii feature.

This change doesn't require any code modifications in jsii-rosetta
because the existing code will handle this feature correctly. We only
need to declare support for it in the SUPPORTED_ASSEMBLY_FEATURES list
to prevent assemblies using this feature from being rejected during
processing.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0