fix(trim-cache): correct log message for trimmed snippet count (#3492)
The log message for `trim-cache` was displaying incorrect information.
It showed `original.count - updated.count` as the remaining count, but
that value is actually the number of trimmed snippets. The message now
correctly shows the trimmed count in parentheses.


### Before

```console
[jsii-rosetta] [INFO] /root/.s3buildcache/rosetta-cache.tabl.json: 3494 snippets remaining (3494 - 3494 trimmed)
```

### After

```console
[jsii-rosetta] [INFO] /root/.s3buildcache/rosetta-cache.tabl.json: 3494 snippets remaining (0 trimmed)
```

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0